Update typehinting for models/graph

+++ deeprank_gnn/models/graph.py
@@ -6,7 +6,7 @@ from networkx import Graph as NetworkxGraph
 class Graph(NetworkxGraph):
     "this is a graph just like in networkx, but with an id and associated target values"
 
-    def __init__(self, id_: str = None, targets: Dict[str, int] = None):
+    def __init__(self, id_: str = None, targets: Dict[str, float] = None):
         """
             Args:
                 id_(str, optional): unique identifier for this graph, random by default
